Move general reflect logic to internal/value
This internal package has the API:
	Format(reflect.Value, useStringer bool) string
	SortKeys(vs []reflect.Value) []reflect.Value

While moving Format and SortKeys, we also tweak the formatting logic
to print the type on named primitive types.
